♻️ refactor: refactor the jwt code (#4322)
* ♻️ refactor: refactor the jwt code

* clean

* fix lint
