+++ src/app/(backend)/middleware/auth/index.test.ts
@@ -1,12 +1,12 @@
-import { getAuth } from '@clerk/nextjs/server';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
 import { AgentRuntimeError } from '@/libs/agent-runtime';
 import { ChatErrorType } from '@/types/fetch';
 import { createErrorResponse } from '@/utils/errorResponse';
+import { getJWTPayload } from '@/utils/server/jwt';
 
 import { RequestHandler, checkAuth } from './index';
-import { checkAuthMethod, getJWTPayload } from './utils';
+import { checkAuthMethod } from './utils';
 
 vi.mock('@clerk/nextjs/server', () => ({
   getAuth: vi.fn(),
@@ -18,6 +18,9 @@ vi.mock('@/utils/errorResponse', () => ({
 
 vi.mock('./utils', () => ({
   checkAuthMethod: vi.fn(),
+}));
+
+vi.mock('@/utils/server/jwt', () => ({
   getJWTPayload: vi.fn(),
 }));
