+++ src/libs/trpc/middleware/jwtPayload.test.ts
@@ -2,10 +2,10 @@
 import { TRPCError } from '@trpc/server';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
-import * as utils from '@/app/(backend)/middleware/auth/utils';
 import { createCallerFactory } from '@/libs/trpc';
 import { trpc } from '@/libs/trpc/init';
 import { AuthContext, createContextInner } from '@/server/context';
+import * as utils from '@/utils/server/jwt';
 
 import { jwtPayloadChecker } from './jwtPayload';
