+++ src/libs/trpc/middleware/jwtPayload.ts
@@ -1,7 +1,7 @@
 import { TRPCError } from '@trpc/server';
 
-import { getJWTPayload } from '@/app/(backend)/middleware/auth/utils';
 import { trpc } from '@/libs/trpc/init';
+import { getJWTPayload } from '@/utils/server/jwt';
 
 export const jwtPayloadChecker = trpc.middleware(async (opts) => {
   const { ctx } = opts;
